Rename compile_gpdb_{custom_config=>open_source} in Master and PR pipelines
Currently, we have a job compile_gpdb_custom_config_centos6 in
pr_pipeline and master that compiles gpdb with a custom config, but the
purpose of this job is unclear. It turns out that this job is for
compiling gpdb the open-source way, so we renamed the job as
compile_gpdb_open_source and changed the name of task yml and scripts.

Signed-off-by: David Sharp <dsharp@pivotal.io>